move room databases to core
