+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/data/mapping/StorageModelMapping.kt
@@ -1,34 +1,9 @@
-package com.github.klee0kai.thekey.app.data.room.entry
+package com.github.klee0kai.thekey.app.data.mapping
 
-import android.os.Parcelable
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.github.klee0kai.thekey.app.data.room.dao.StorageFilesDao
 import com.github.klee0kai.thekey.app.domain.model.ColoredStorage
 import com.github.klee0kai.thekey.app.engine.model.Storage
+import com.github.klee0kai.thekey.core.data.room.entry.StorageFileEntry
 import com.github.klee0kai.thekey.core.domain.ColorGroup
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-@Entity(tableName = StorageFilesDao.TABLE_NAME)
-data class StorageFileEntry(
-    @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(index = true, name = "id")
-    val id: Long = 0,
-
-    @ColumnInfo(name = "path")
-    val path: String = "",
-
-    @ColumnInfo(name = "name")
-    val name: String = "",
-
-    @ColumnInfo(name = "description")
-    val description: String = "",
-
-    @ColumnInfo(name = "color_group")
-    val coloredGroupId: Long = 0,
-) : Parcelable
 
 
 fun StorageFileEntry.toStorage(): Storage = this.run {
@@ -57,4 +32,3 @@ fun ColoredStorage.toStorageEntry(
         coloredGroupId = colorGroup?.id ?: 0
     )
 }
-
