+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/dependencies/AppComponentProviders.kt
@@ -6,5 +6,4 @@ interface AppComponentProviders :
     HelperDependencies,
     InteractorsDependencies,
     RepositoriesDependencies,
-    EngineDependencies,
-    DBDependencies
+    EngineDependencies
