+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/dependencies/RepositoriesDependencies.kt
@@ -1,6 +1,5 @@
 package com.github.klee0kai.thekey.app.di.dependencies
 
-import com.github.klee0kai.thekey.app.data.repositories.settings.SettingsRepository
 import com.github.klee0kai.thekey.app.data.repositories.storage.GroupsRepository
 import com.github.klee0kai.thekey.app.data.repositories.storage.LoginnedStorages
 import com.github.klee0kai.thekey.app.data.repositories.storage.NotesRepository
@@ -12,8 +11,6 @@ import com.github.klee0kai.thekey.core.di.wrap.AsyncCoroutineProvide
 interface RepositoriesDependencies {
 
     fun storagesRepositoryLazy(): AsyncCoroutineProvide<StoragesRepository>
-
-    fun settingsRepositoryLazy(): AsyncCoroutineProvide<SettingsRepository>
 
     fun loginedRepLazy(): AsyncCoroutineProvide<LoginnedStorages>
