+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/di/modules/RepositoriesModule.kt
@@ -2,7 +2,6 @@ package com.github.klee0kai.thekey.app.di.modules
 
 import com.github.klee0kai.stone.annotations.module.Module
 import com.github.klee0kai.stone.annotations.module.Provide
-import com.github.klee0kai.thekey.app.data.repositories.settings.SettingsRepository
 import com.github.klee0kai.thekey.app.data.repositories.storage.GroupsRepository
 import com.github.klee0kai.thekey.app.data.repositories.storage.LoginnedStorages
 import com.github.klee0kai.thekey.app.data.repositories.storage.NotesRepository
@@ -16,9 +15,6 @@ interface RepositoriesModule {
 
     @Provide(cache = Provide.CacheType.Soft)
     fun storagesRepository(): StoragesRepository = StoragesRepositoryImpl()
-
-    @Provide(cache = Provide.CacheType.Soft)
-    fun settingsRepository(): SettingsRepository
 
     @Provide(cache = Provide.CacheType.Strong)
     fun loginnedStorages(): LoginnedStorages
