+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/data/repository/settings/delegates/BooleanNoteDelegate.kt
@@ -1,6 +1,6 @@
-package com.github.klee0kai.thekey.app.data.repositories.settings.delegates
+package com.github.klee0kai.thekey.core.data.repository.settings.delegates
 
-import com.github.klee0kai.thekey.app.data.room.dao.SettingDao
+import com.github.klee0kai.thekey.core.data.room.dao.SettingDao
 import com.github.klee0kai.thekey.core.di.wrap.AsyncCoroutineProvide
 import kotlinx.coroutines.CoroutineScope
