+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/data/room/dao/SettingDao.kt
@@ -1,10 +1,10 @@
-package com.github.klee0kai.thekey.app.data.room.dao
+package com.github.klee0kai.thekey.core.data.room.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.github.klee0kai.thekey.app.data.room.entry.SettingPairEntry
+import com.github.klee0kai.thekey.core.data.room.entry.SettingPairEntry
 
 @Dao
 interface SettingDao {
