+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/data/room/entry/SettingPairEntry.kt
@@ -1,10 +1,10 @@
-package com.github.klee0kai.thekey.app.data.room.entry
+package com.github.klee0kai.thekey.core.data.room.entry
 
 import android.os.Parcelable
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.github.klee0kai.thekey.app.data.room.dao.SettingDao
+import com.github.klee0kai.thekey.core.data.room.dao.SettingDao
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
