+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/di/dependecies/CoreDBDependencies.kt
@@ -1,0 +1,19 @@
+package com.github.klee0kai.thekey.core.di.dependecies
+
+import com.github.klee0kai.thekey.core.data.room.KeyDatabase
+import com.github.klee0kai.thekey.core.data.room.dao.ColorGroupDao
+import com.github.klee0kai.thekey.core.data.room.dao.SettingDao
+import com.github.klee0kai.thekey.core.data.room.dao.StorageFilesDao
+import com.github.klee0kai.thekey.core.di.wrap.AsyncCoroutineProvide
+
+interface CoreDBDependencies {
+
+    fun dbLazy(): AsyncCoroutineProvide<KeyDatabase>
+
+    fun settingDaoLazy(): AsyncCoroutineProvide<SettingDao>
+
+    fun storageDaoLazy(): AsyncCoroutineProvide<StorageFilesDao>
+
+    fun colorGroupDaoLazy(): AsyncCoroutineProvide<ColorGroupDao>
+
+}
