+++ core/src/main/kotlin/com/github/klee0kai/thekey/core/di/dependecies/CoreRepositoryDependencies.kt
@@ -1,0 +1,10 @@
+package com.github.klee0kai.thekey.core.di.dependecies
+
+import com.github.klee0kai.thekey.core.data.repository.settings.SettingsRepository
+import com.github.klee0kai.thekey.core.di.wrap.AsyncCoroutineProvide
+
+interface CoreRepositoryDependencies {
+
+    fun settingsRepositoryLazy(): AsyncCoroutineProvide<SettingsRepository>
+
+}
